app/requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
crispy-bootstrap5==2024.2
cryptography==38.0.4
daphne==4.0.0
dj-database-url==2.2.0
Django==4.2.13
django-crispy-forms==2.1
djangorestframework==3.15.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyMySQL==1.1.1
pyOpenSSL==22.1.0
pytz==2024.1
requests==2.32.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
txaio==23.1.1
typing_extensions==4.12.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
whitenoise==6.6.0
zope.interface==6.4.post2
